+++ app/src/main/java/ryey/easer/core/data/storage/ProfileDataStorage.java
@@ -2,25 +2,26 @@ package ryey.easer.core.data.storage;
 
 import android.content.Context;
 
-import java.io.IOException;
 import java.util.List;
 
 import ryey.easer.core.data.ProfileStructure;
-import ryey.easer.core.data.storage.json.profile.JsonProfileDataStorageBackend;
-import ryey.easer.core.data.storage.xml.profile.XmlProfileDataStorageBackend;
+import ryey.easer.core.data.storage.backend.json.profile.JsonProfileDataStorageBackend;
+import ryey.easer.core.data.storage.backend.xml.profile.XmlProfileDataStorageBackend;
 
 public class ProfileDataStorage {
 
     private static ProfileDataStorage instance = null;
 
+    Context context;
     JsonProfileDataStorageBackend json_storage;
     XmlProfileDataStorageBackend xml_storage;
 
-    public static ProfileDataStorage getInstance(Context context) throws IOException {
+    public static ProfileDataStorage getInstance(Context context) {
         if (instance == null) {
             instance = new ProfileDataStorage();
             instance.json_storage = JsonProfileDataStorageBackend.getInstance(context);
             instance.xml_storage = XmlProfileDataStorageBackend.getInstance(context);
+            instance.context = context;
         }
         return instance;
     }
@@ -45,6 +46,8 @@ public class ProfileDataStorage {
     }
 
     public boolean delete(String name) {
+        if (!StorageHelper.isSafeToDeleteProfile(context, name))
+            return false;
         if (json_storage.list().contains(name))
             return json_storage.delete(name);
         else if (xml_storage.list().contains(name))
@@ -53,12 +56,36 @@ public class ProfileDataStorage {
     }
 
     public boolean edit(String oldName, ProfileStructure profile) {
+        if (oldName.equals(profile.getName()))
+            return update(profile);
         if (!add(profile))
             return false;
+        boolean succees;
         if (json_storage.list().contains(oldName))
-            return json_storage.delete(oldName);
+            succees = json_storage.delete(oldName);
         else if (xml_storage.list().contains(oldName))
-            return xml_storage.delete(oldName);
+            succees = xml_storage.delete(oldName);
+        else {
+            throw new IllegalAccessError();
+        }
+        if (succees && oldName.equals(profile.getName())) {
+            EventDataStorage eventDataStorage = EventDataStorage.getInstance(context);
+            succees = eventDataStorage.handleProfileRename(oldName, profile.getName());
+        }
+        return succees;
+    }
+
+    boolean update(ProfileStructure profileStructure) {
+        String name = profileStructure.getName();
+        boolean success;
+        if (json_storage.list().contains(name))
+            success = json_storage.delete(name);
+        else if (xml_storage.list().contains(name))
+            success = xml_storage.delete(name);
+        else
+            throw new IllegalAccessError();
+        if (success)
+            return add(profileStructure);
         return false;
     }
 }
